Move testers to lib project
