+++ lib/build.gradle
@@ -43,4 +43,6 @@ dependencies {
     compile "openrio.mirror.third.kauailabs:navx-java:3.0.346"
     compile "jaci.pathfinder:Pathfinder-Java:1.8"
     compile "openrio.powerup:MatchData:2018.01.07"
+    compile "com.google.guava:guava:27.0-jre"
+    compile "org.apache.commons:commons-math3:3.6.1"
 }
