+++ lib/src/main/java/org/team1540/base/testers/AbstractTester.java
@@ -1,4 +1,4 @@
-package org.team1540.base.testing;
+package org.team1540.base.testers;
 
 import com.google.common.collect.EvictingQueue;
 import java.util.Collections;
