+++ lib/src/main/java/org/team1540/base/testers/ResultWithMetadata.java
@@ -1,4 +1,4 @@
-package org.team1540.base.testing;
+package org.team1540.base.testers;
 
 import org.jetbrains.annotations.NotNull;
