+++ test/build.gradle
@@ -16,8 +16,6 @@ dependencies {
     compile ctre()
     compile navx()
     compile pathfinder()
-    compile 'com.google.guava:guava:27.0-jre'
-    compile 'org.apache.commons:commons-math3:3.6.1'
 }
 
 def TEAM = 1540
